+++ src/test/groovy/BeatsWhenStepTests.groovy
@@ -302,6 +302,149 @@ class BeatsWhenStepTests extends ApmBasePipelineTest {
   }
 
   @Test
+  void test_whenNotChangeset_and_no_data() throws Exception {
+    def script = loadScript(scriptName)
+    def ret = script.whenNotChangeset()
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
+  void test_whenNotChangeset_and_content_without_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'Jenkinsfile'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangeset(content: [ not_changeset: ['^.ci']])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangeset_and_content_with_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'Jenkinsfile'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangeset(content: [ not_changeset: ['^Jenkinsfile']])
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
+  void test_whenNotChangeset_content_and_macro() throws Exception {
+    def script = loadScript(scriptName)
+    def ret = script.whenNotChangeset(content: [ not_changeset: ['^.ci', '@oss']],
+                                      changeset: [ oss: [ '^oss'] ])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangeset_content_and_macro_with_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'oss'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangeset(content: [ not_changeset: ['^.ci', '@oss']],
+                                      changeset: [ oss: [ '^oss'] ])
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
+  void test_whenNotChangeset_content_and_macro_without_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'oss'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangeset(content: [ not_changeset: ['^.ci', '@osss']],
+                                      changeset: [ oss: [ '^oss'] ])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_and_no_data() throws Exception {
+    def script = loadScript(scriptName)
+    def ret = script.whenNotChangesetFullMatch()
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_and_content_without_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'Jenkinsfile'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '^.ci'])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_and_content_with_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'Jenkinsfile'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '^Jenkinsfile'])
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_and_content_with_full_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = '''.ci/Jenkinsfile
+.ci/jobs
+'''
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '^.ci.*'])
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_and_content_without_full_match_no_matches() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = '''.ci/Jenkinsfile
+.ci/jobs
+'''
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '^.foo.*'])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_and_content_without_full_match_with_partial_matches() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = '''.ci/Jenkinsfile
+.foo/jobs
+'''
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '^.ci.*'])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_content_and_macro() throws Exception {
+    def script = loadScript(scriptName)
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '@oss'],
+                                               changeset: [ oss: [ '^oss'] ])
+    printCallStack()
+    assertTrue(ret)
+  }
+
+  @Test
+  void test_whenNotChangesetFullMatch_content_and_macro_with_match() throws Exception {
+    def script = loadScript(scriptName)
+    def changeset = 'oss'
+    helper.registerAllowedMethod('readFile', [String.class], { return changeset })
+    def ret = script.whenNotChangesetFullMatch(content: [ not_changeset_full_match: '@oss'],
+                                               changeset: [ oss: [ '^oss'] ])
+    printCallStack()
+    assertFalse(ret)
+  }
+
+  @Test
   void test_whenParameters_and_no_params() throws Exception {
     def script = loadScript(scriptName)
     def ret = script.whenParameters()